JAMES-3589 Move mailetcontainer classes out of camel package
