+++ server/mailet/mailetcontainer-camel/src/main/java/org/apache/james/mailetcontainer/impl/CamelCompositeProcessor.java
@@ -17,7 +17,7 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.mailetcontainer.impl.camel;
+package org.apache.james.mailetcontainer.impl;
 
 import javax.annotation.PostConstruct;
 import javax.inject.Inject;
