+++ server/mailet/mailetcontainer-camel/src/main/java/org/apache/james/mailetcontainer/lib/AbstractStateMailetProcessor.java
@@ -38,6 +38,7 @@ import org.apache.james.lifecycle.api.Configurable;
 import org.apache.james.mailetcontainer.api.MailProcessor;
 import org.apache.james.mailetcontainer.api.MailetLoader;
 import org.apache.james.mailetcontainer.api.MatcherLoader;
+import org.apache.james.mailetcontainer.impl.CamelProcessor;
 import org.apache.james.mailetcontainer.impl.MailetConfigImpl;
 import org.apache.james.mailetcontainer.impl.MatcherConfigImpl;
 import org.apache.james.mailetcontainer.impl.MatcherMailetPair;
@@ -344,7 +345,7 @@ public abstract class AbstractStateMailetProcessor implements MailProcessor, Con
 
     /**
      * Setup the routing for the configured {@link MatcherMailetPair}'s for this
-     * {@link org.apache.james.mailetcontainer.impl.camel.CamelProcessor}
+     * {@link CamelProcessor}
      */
     protected abstract void setupRouting(List<MatcherMailetPair> pairs) throws MessagingException;
